[REFACTOR] service search/index로 분리
