+++ src/main/java/com/merseongsanghoe/sooljarisearchengine/service/SearchService.java
@@ -1,16 +1,14 @@
 package com.merseongsanghoe.sooljarisearchengine.service;
 
 import com.merseongsanghoe.sooljarisearchengine.DAO.AlcoholElasticsearchRepository;
-import com.merseongsanghoe.sooljarisearchengine.DAO.AlcoholRepository;
 import com.merseongsanghoe.sooljarisearchengine.DTO.SearchResultDTO;
-import com.merseongsanghoe.sooljarisearchengine.domain.*;
+import com.merseongsanghoe.sooljarisearchengine.domain.AlcoholDocument;
+import com.merseongsanghoe.sooljarisearchengine.domain.TagDocument;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.elasticsearch.core.SearchHit;
 import org.springframework.data.elasticsearch.core.SearchHits;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -18,11 +16,9 @@ import java.util.List;
 import java.util.Map;
 
 @Service
-@Slf4j
 @RequiredArgsConstructor
-public class AlcoholService {
+public class SearchService {
 
-    private final AlcoholRepository alcoholRepository;
     private final AlcoholElasticsearchRepository alcoholElasticsearchRepository;
 
     /**
@@ -52,7 +48,6 @@ public class AlcoholService {
                     .degree(hit.getContent().getDegree())
                     .build();
 
-            // TODO: 리스트 자료형도 빌더 패턴에 포함할 수 있을까? - 1
             for (TagDocument tag : hit.getContent().getTags()) {
                 dto.addTag(tag.getTitle());
             }
@@ -62,34 +57,5 @@ public class AlcoholService {
         result.put("data", searchResultDTOList);
 
         return result;
-    }
-
-    /**
-     * 데이터베이스 속 alcohol 데이터 전부 elasticsearch에 인덱싱
-     */
-    @Transactional(readOnly = true)
-    public void indexAll() {
-        // TODO: 데이터가 많을 때 전체 인덱싱 성능 확인 필요
-        List<Alcohol> alcohols = alcoholRepository.findAllWithSearchKeys();
-
-        for (Alcohol alcohol : alcohols) {
-            // alcohol 엔티티를 인덱싱할 형태로 변환
-            AlcoholDocument alcoholDocument = AlcoholDocument.builder()
-                    .alcoholId(alcohol.getId())
-                    .title(alcohol.getTitle())
-                    .category(alcohol.getCategory())
-                    .degree(alcohol.getDegree())
-                    .build();
-
-            // TODO: 리스트 자료형도 빌더 패턴에 포함할 수 있을까? - 2
-            for (Tag tag : alcohol.getTags()) {
-                alcoholDocument.addTag(new TagDocument(tag.getTitle()));
-            }
-            for (AlcSearchKey searchKey : alcohol.getSearchKeys()) {
-                alcoholDocument.addSearchKey(new AlcSearchKeyDocument(searchKey.getKey()));
-            }
-
-            alcoholElasticsearchRepository.save(alcoholDocument);
-        }
     }
 }
